lca-tv-website/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0